Switched from qmake to cmake build system and added some clean-up and small re-estructuring

+++ ui/dialogs/exportEsquema_dlg.cpp
@@ -1,9 +1,15 @@
+// =================================================== \\
+// ====     Copyright (c) 2024 Ignasi Camps       ==== \\
+// ==== SPDX-License-Identifier: GPL-3.0-or-later ==== \\
+// =================================================== \\
+
 #include "exportEsquema_dlg.h"
 #include "ui_exportEsquema_dlg.h"
 
 #include "document/CMDoc.h"
+#include "document/CEsquemaDoc.h"
 
-exportEsquema_dlg::exportEsquema_dlg(QWidget *parent, int* pSelectedIndex)
+exportEsquema_dlg::exportEsquema_dlg(QWidget *parent, size_t *pSelectedIndex)
     : QDialog(parent), ui(new Ui::exportEsquema_dlg), m_selectedIndex{pSelectedIndex} {
     ui->setupUi(this);
