+++ ui/dialogs/exportEsquema_dlg.h
@@ -1,3 +1,8 @@
+// =================================================== \\
+// ====     Copyright (c) 2024 Ignasi Camps       ==== \\
+// ==== SPDX-License-Identifier: GPL-3.0-or-later ==== \\
+// =================================================== \\
+
 #ifndef EXPORTESQUEMA_DLG_H
 #define EXPORTESQUEMA_DLG_H
 
@@ -12,7 +17,7 @@ class exportEsquema_dlg : public QDialog
     Q_OBJECT
 
 public:
-    explicit exportEsquema_dlg(QWidget *parent, int *pSelectedIndex);
+    explicit exportEsquema_dlg(QWidget *parent, size_t *pSelectedIndex);
     ~exportEsquema_dlg();
 
 private slots:
@@ -20,7 +25,7 @@ private slots:
 
 private:
     Ui::exportEsquema_dlg *ui;
-    int *m_selectedIndex;
+    size_t *m_selectedIndex;
 };
 
 #endif // EXPORTESQUEMA_DLG_H
